chore: restore auth file and move the server action to user file
