+++ ui/actions/auth/auth.ts
@@ -1,12 +1,9 @@
 "use server";
 
-import { revalidatePath } from "next/cache";
 import { AuthError } from "next-auth";
 import { z } from "zod";
 
 import { signIn, signOut } from "@/auth.config";
-import { auth } from "@/auth.config";
-import { parseStringify } from "@/lib";
 import { authFormSchema } from "@/types";
 
 const formSchemaSignIn = authFormSchema("sign-in");
@@ -139,34 +136,6 @@ export const getToken = async (formData: z.infer<typeof formSchemaSignIn>) => {
     };
   } catch (error) {
     throw new Error("Error in trying to get token");
-  }
-};
-
-export const getProfileInfo = async () => {
-  const session = await auth();
-  const keyServer = process.env.API_BASE_URL;
-  const url = new URL(`${keyServer}/users/me`);
-
-  try {
-    const response = await fetch(url.toString(), {
-      method: "GET",
-      headers: {
-        Accept: "application/vnd.api+json",
-        Authorization: `Bearer ${session?.accessToken}`,
-      },
-    });
-
-    if (!response.ok) {
-      throw new Error(`Failed to fetch user data: ${response.statusText}`);
-    }
-
-    const data = await response.json();
-    const parsedData = parseStringify(data);
-    revalidatePath("/profile");
-    return parsedData;
-  } catch (error) {
-    console.error("Error fetching profile:", error);
-    return undefined;
   }
 };
